Moved auth and import queue services to `sharedClient`
